TP1-Java

+++ Java II/Aula4.1-Ex_II_Operacao_com_Matrizes/src/main/java/org/example/Calculadora.java
@@ -4,17 +4,17 @@ import javax.swing.*;
 
 public class Calculadora {
 
-    public double somar(double num1, double num2) {
+    public int somar(int num1, int num2) {
         return num1 + num2;
     }
 
-    public double subtrair(double num1, double num2) {
+    public int subtrair(int num1, int num2) {
         return num1 - num2;
     }
 
-    public double multiplicar(double num1, double num2) {
+    public int multiplicar(int num1, int num2) {
         return num1 * num2;
     }
 
-    //public double escalar(double num, double escalar) { return num * escalar; }
+    //public int escalar(int num, int escalar) { return num * escalar; }
 }
